+++ yarn-project/p2p/src/services/reqresp/protocols/block_txs/index.ts
@@ -1,0 +1,3 @@
+export * from './block_txs_handler.js';
+export * from './bitvector.js';
+export * from './block_txs_reqresp.js';
