+++ yarn-project/p2p/src/services/reqresp/protocols/index.ts
@@ -6,6 +6,4 @@ export * from './status.js';
 export * from './tx.js';
 export * from './goodbye.js';
 export * from './block.js';
-export * from './auth.js';
-export * from './block_txs.js';
-export * from './block_txs_handler.js';
+export * from './block_txs/index.js';
